処理を分割。
